Refatoring OOP structure
